Split the file for the different annealing solvers into individual files.
